+++ changelog.md
@@ -2,9 +2,15 @@
 
 ## 2.2.0
 
+Note: this version includes networking changes, servers and clients must be upgraded together.
+
+- Added an option to reverse the default transfer iteration order
 - Added an option to arrange buttons horizontally rather than vertically
 - Added support for creating separate class policies for unique inventory titles
 - Added a per-policy option to anchor buttons to the reference slot instead of the GUI edge
+
+## 2.1.8
+
 - Fixed detection of CTRL/CMD keys in 1.21.9-11
 
 ## 2.1.7